+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 - `forks` block function [10ab025](https://github.com/gabiseabra/redux-saga-test-factory/commit/10ab025)
 - typescript typings [2cc9d5a](https://github.com/gabiseabra/redux-saga-test-factory/commit/2cc9d5a)
+- `clone` can be used as a block function
 
 ## v0.1.1